# spec/swagger.yaml
swagger: '2.0'
info:
  title: SuggestGrid
  version: 0.1.32-SNAPSHOT
  description: Personalization made Simple
  x-codegen-settings:
    appendContentHeaders: true
    generateAsyncCode: true
    useMethodPrefix: false
    useModelPostfix: false
    useEnumPostfix: false
    useConstructorsForConfig: false
    iOSUseAppInfoPlist: true
    androidUseAppManifest: true
    collectParameters: false
    csharpDefaultNamespace: SuggestGrid
    javaDefaultPackageName: com.suggestgrid
    brandLabel: SuggestGrid
    userAgent: SUGGESTGRID
    projectName: SuggestGrid
    enableAdditionalModelProperties: false
    preserveParameterOrder: true
    enablePHPComposerVersionString: true
schemes:
  - https
securityDefinitions:
  basicAuth:
    type: basic
    description: HTTP Basic Authentication.
security:
  - basicAuth: []
consumes:
  - application/json
produces:
  - application/json
...
definitions:
  Action:
    type: object
    description: |
      An action object represents a user's action of a type on an item.
    properties:
      type:
        type: string
        description: The type that the action belongs to.
      user_id:
        type: string
        description: The user id of the performer of the action.
      item_id:
        type: string
        description: The item id of the item the action is performed on.
      rating:
        type: number
        description: The optional rating given by the user, if the type is explicit.
      timestamp:
        type: integer
        format: int64
        description: The optional UNIX epoch timestamp of the action. Defaults to the current timestamp.
    required:
      - type
      - user_id
      - item_id
...
  MessageResponse:
    type: object
    description: Basic response with a message for success responses.
    properties:
      message:
        type: string
        description: Message of the response.
  ActionsResponse:
    type: object
    description: Paged actions response.
    properties:
      count:
        type: integer
        format: int64
        description: The number of actions in the response.
      total_count:
        type: integer
        format: int64
        description: The total number of actions.
      actions:
        type: array
        items:
          $ref: '#/definitions/Action'
  ErrorResponse:
    type: object
    description: Error response.
    properties:
      error_text:
        type: string
        description: Message of the response.
      error_description:
        type: string
        description: Description of the response.
      error_uri:
        type: string
        description: The URI of the error for more details.
...
  DeleteErrorResponse:
    type: object
    description: Error response with found, deleted, and failed parameters.
    properties:
      error_text:
        type: string
        description: Message of the response.
      error_description:
        type: string
        description: Description of the response.
      error_uri:
        type: string
        description: The URI of the error for more details.
      found:
        type: integer
        format: int64
        description: The number of records found.
      deleted:
        type: integer
        format: int64
        description: The number of records deleted.
      failed:
        type: integer
        format: int64
        description: The number of records failed to be deleted.
  DeleteSuccessResponse:
    type: object
    description: Successful delete response with found, deleted, and failed parameters.
    properties:
      message:
        type: string
        description: Message of the response.
      found:
        type: integer
        format: int64
        description: The number of records found.
      deleted:
        type: integer
        format: int64
        description: The number of records deleted.
      failed:
        type: integer
        format: int64
        description: The number of records failed to be deleted.
...
  BulkPostResponse:
    type: object
    description: |
      Bulk error response.
      Returned on all bulk post requests regardless of the number of erroneous or successful posts.
    properties:
      message:
        type: string
        description: Message of the response.
      errors:
        type: array
        items:
          $ref: '#/definitions/BulkPostError'
  BulkPostError:
    type: object
    description: Error response returned due to bulk post schema validations.
    properties:
      message:
        type: string
        description: Message of the response.
      value:
        type: object
        description: The cause of the error.
      error:
        type: object
        description: Programatic description of the error.
...
paths:
...
  /v1/actions:
    post:
      tags:
        - action
      operationId: post_action
      summary: Posts an Action
      description: |
        Posts an action to the given type in the body.
        The body must have user id, item id and type.
        Rating is required for actions sent to an explicit type.
      parameters:
        - name: action
          in: body
          description: The action to be posted.
          required: true
          schema:
            $ref: '#/definitions/Action'
      responses:
        201:
          description: An action is created.
          schema:
            $ref: '#/definitions/MessageResponse'
        400:
          description: Required user id or item id parameters are missing from the request.
          schema:
            $ref: '#/definitions/ErrorResponse'
        402:
          description: Action limit exceeded.
          schema:
            $ref: '#/definitions/ErrorResponse'
        404:
          description: Action type does not exists.
          schema:
            $ref: '#/definitions/ErrorResponse'
        default:
          description: Unexpected internal error.
          schema:
            $ref: '#/definitions/ErrorResponse'
    get:
      tags:
        - action
      operationId: get_actions
      summary: Gets Actions
      description: |
        Get actions. Defaut responses will be paged by 10 actios each.
        Type, user id, item id, or older than parameters could be provided.
        The intersection of the provided parameters will be returned.
      parameters:
        - name: type
          in: query
          description: Get actions of a type.
          required: false
          type: string
          format: id
        - name: user_id
          in: query
          description: Get actions of a user id.
          required: false
          type: string
          format: id
        - name: item_id
          in: query
          description: Get actions of an item id.
          required: false
          type: string
          format: id
        - name: older_than
          in: query
          description: |
            Get actions older than the given duration, or the given time number.
            Could be a ISO 8601 duration, or a Unix time number.
            Specifications are available at https://en.wikipedia.org/wiki/ISO_8601#Durations, or https://en.wikipedia.org/wiki/Unix_time.
          required: false
          type: string
        - name: size
          in: query
          description: |
            The number of the users response.
            Defaults to 10. Must be between 1 and 10,000 inclusive.
            This parameter must be string represetation of an integer like "1".
          required: false
          type: integer
          format: int64
        - name: from
          in: query
          description: |
            The number of users to be skipped from the response.
            Defaults to 0. Must be bigger than or equal to 0.
            This parameter must be string represetation of an integer like "1".
          required: false
          type: integer
          format: int64
      responses:
        200:
          description: Paged actions response.
          schema:
            $ref: '#/definitions/ActionsResponse'
        default:
          description: Unexpected internal error.
          schema:
            $ref: '#/definitions/ErrorResponse'
    delete:
      tags:
        - action
      operationId: delete_actions
      summary: Delete Actions
      description: |
        Warning: Please use get actions with the exact parameters first to inspect the actions to be deleted.
        
        * Type must be provided. 
        * If user id is provided, all actions of the user will be deleted.
        * If item id is provided, all actions on the item will be deleted.
        * If older than is provided, all actions older than the timestamp or the duration will be deleted.
        * If a number of these parameters are provided, the intersection of these parameters will be deleted.
        * In addition to a type, at least one of these parameters must be provided. In order to delete all the actions of a type, delete the type.
      parameters:
        - name: type
          in: query
          description: Delete actions of a type. This parameter and at least one other parameter is required.
          required: true
          type: string
          format: id
        - name: user_id
          in: query
          description: Delete actions of a user id.
          required: false
          type: string
          format: id
        - name: item_id
          in: query
          description: Delete actions of an item id.
          required: false
          type: string
          format: id
        - name: older_than
          in: query
          description: |
            Delete actions older than the given duration, or the given time number.
            Could be a ISO 8601 duration, or a Unix time number.
            Specifications are available at https://en.wikipedia.org/wiki/ISO_8601#Durations, or https://en.wikipedia.org/wiki/Unix_time.
          required: false
          type: string
      responses:
        200:
          description: Successfully deleted actions.
          schema:
            $ref: '#/definitions/DeleteSuccessResponse'
        400:
          description: Required user id or item id parameters are missing.
          schema:
            $ref: '#/definitions/ErrorResponse'
        404:
          description: Delete actions not found.
          schema:
            $ref: '#/definitions/DeleteErrorResponse'
        422:
          description: No query parameter (user id, item id, or older than) is given. qIn order to delete all actionsdelete the type.
          schema:
            $ref: '#/definitions/ErrorResponse'
        505:
          description: Request timed out.
          schema:
            $ref: '#/definitions/ErrorResponse'
        default:
          description: Unexpected internal error.
          schema:
            $ref: '#/definitions/ErrorResponse'
  /v1/actions/_bulk:
    post:
      tags:
        - action
      operationId: post_bulk_actions
      summary: Posts Actions
      description: |
        Posts bulk actions to SuggestGrid.
        The recommended method for posting multiple actions at once.
      consumes:
       - text/plain; charset=utf-8
      parameters:
        - name: actions
          in: body
          description: |
            A number of action objects separated with newlines.
            Note that this is not a valid JSON data structure.
            The body size is limited to 10 thousand lines.
          required: true
          schema:
            type: string
      responses:
        201:
          description: Bulk action request executed.
          schema:
            $ref: '#/definitions/BulkPostResponse'
        400:
          description: Body is missing.
          schema:
            $ref: '#/definitions/ErrorResponse'
        402:
          description: Action limit exceeded.
          schema:
            $ref: '#/definitions/ErrorResponse'
        404:
          description: Action type does not exists.
          schema:
            $ref: '#/definitions/ErrorResponse'
        413:
          description: Bulk request maximum line count exceeded.
          schema:
            $ref: '#/definitions/ErrorResponse'
        default:
          description: Unexpected internal error.
          schema:
            $ref: '#/definitions/ErrorResponse'
  # Metadata Paths
...